Add option for Vec4 value types
This allows you to download/upload arrays of Vec4 directly, skipping the
packing/unpacking phase for floats. It might be faster and more precise. For
example:

    monkeys.set("a", [0, 1, 2, 3]);
    monkeys.work(1, "a(0) := a(0) + vec4(0.0, 1.0, 2.0, 3.0);");
    console.log(monkeys.get("a"));

This outputs [0, 2, 4, 6]. Notice the range of each component is [0..255], each
that monkey index refers to 4 numbers of the array in a row (since it is a Vec4).